Consolidate some dtos
